tweak in assembly equality comparer
Sometimes I got strange errors from the MemoizedFunction, stating that
an object with same key has already been added to the arguments cache,
but this shouldn't be possible by looking at the Invoke code. I
suspected this happens for dynamically generated assemblies and by
heuristics concluded that there might be a problem with equality
comparer used with memoized function. Modified the
assemblyequalitycomparer class, the change should not break anything.